move Message type into CRDT package
