+++ packages/crdt/src/main.ts
@@ -11,3 +11,12 @@ export {
 } from './crdt';
 
 export const SyncProtoBuf = SyncPb;
+
+export type Message = {
+  column: string;
+  dataset: string;
+  old?: unknown;
+  row: string;
+  timestamp: string;
+  value: string | number | null;
+};
